requirements.txt
flask
numpy
onnx
onnxruntime
pillow
